# api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "bloggorras Api Documentation",
    "description": "bloggorras Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "bloggorras Logo",
      "altText": "bloggorras Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "bloggorras API Support",
      "url": "https://bloggorras.com/info/support",
      "email": "support@bloggorras.com"
    },
    "termsOfService": "https://bloggorras.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/productos": {
      "get": {
        "summary": "List Productos",
        "description": "Productos",
        "operationId": "ListProductos",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProductos"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/productos/view/{id}": {
      "get": {
        "summary": "View Productos",
        "description": "Ver",
        "operationId": "ViewProductos",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewProductos"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/productos/add": {
      "post": {
        "summary": "Add Productos",
        "description": "Agregar nuevo",
        "operationId": "AddProductos",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Productos"
        ],
        
        "requestBody": {
          "description": "Add Productos",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProductos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewProductos"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/productos/edit/{id}": {
      "put": {
        "summary": "Edit Productos",
        "description": "Editar",
        "operationId": "EditProductos",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "requestBody": {
          "description": "Edit Productos",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditProductos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditProductos"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/productos/delete/{id}": {
      "delete": {
        "summary": "Delete Productos",
        "description": "Productos",
        "operationId": "DeleteProductos",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ListProductos": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "nombre": {
            "description": "Nombre",
            "type": "string"
          },
          "precio": {
            "description": "Precio",
            "type": "string"
          },
          "material": {
            "description": "Material",
            "type": "string"
          },
          "tipo_id": {
            "description": "Tipo Id",
            "type": "string"
          },
          "tallas_id": {
            "description": "Tallas Id",
            "type": "string"
          },
          "proveedor_id": {
            "description": "Proveedor Id",
            "type": "string"
          },
          "image": {
            "description": "Image",
            "type": "string"
          }
        }
      },
      "ViewProductos": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "nombre": {
            "description": "Nombre",
            "type": "string"
          },
          "precio": {
            "description": "Precio",
            "type": "string"
          },
          "material": {
            "description": "Material",
            "type": "string"
          },
          "tipo_id": {
            "description": "Tipo Id",
            "type": "string"
          },
          "tallas_id": {
            "description": "Tallas Id",
            "type": "string"
          },
          "proveedor_id": {
            "description": "Proveedor Id",
            "type": "string"
          },
          "image": {
            "description": "Image",
            "type": "string"
          }
        }
      },
      "AddProductos": {
        "type": "object",
        "properties": {
          "nombre": {
            "description": "Nombre",
            "type": "string"
          },
          "precio": {
            "description": "Precio",
            "type": "string"
          },
          "material": {
            "description": "Material",
            "type": "string"
          },
          "tipo_id": {
            "description": "Tipo Id",
            "type": "string"
          },
          "tallas_id": {
            "description": "Tallas Id",
            "type": "string"
          },
          "proveedor_id": {
            "description": "Proveedor Id",
            "type": "string"
          },
          "image": {
            "description": "Image",
            "type": "string"
          }
        }
      },
      "EditProductos": {
        "type": "object",
        "properties": {
          "nombre": {
            "description": "Nombre",
            "type": "string"
          },
          "precio": {
            "description": "Precio",
            "type": "string"
          },
          "material": {
            "description": "Material",
            "type": "string"
          },
          "tipo_id": {
            "description": "Tipo Id",
            "type": "string"
          },
          "tallas_id": {
            "description": "Tallas Id",
            "type": "string"
          },
          "proveedor_id": {
            "description": "Proveedor Id",
            "type": "string"
          },
          "image": {
            "description": "Image",
            "type": "string"
          }
        }
      },
...
      "SearchQuery": {
        "type": "string",
        "properties": {
          "search": {
            "description": "Search value",
            "type": "string"
          }
        }
      },
...
      "LimitQuery": {
        "type": "string",
        "properties": {
          "limit": {
            "description": "Maximum number of records to select",
            "type": "integer"
          }
        }
      },
      "PageNumberQuery": {
        "type": "string",
        "properties": {
          "pagenumber": {
            "description": "Current page number",
            "type": "integer"
          }
        }
      },
      "RecordID": {
        "type": "string",
        "properties": {
          "id": {
            "description": "Record ID",
            "type": "string"
          }
        }
      },
...
      "UnAuthenticated": {
        "type": "object",
        "properties": {
          "code": {
            "description": "401",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthenticated",
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}